refactor: Restructure property filter unit tests
