+++ src/property-filter/__tests__/property-filter.test.tsx
@@ -2,16 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 import React, { useState } from 'react';
 import { act, render } from '@testing-library/react';
-
-import { i18nStrings } from './common';
-
-import createWrapper, {
-  ElementWrapper,
-  PropertyFilterWrapper,
-  PopoverWrapper,
-} from '../../../lib/components/test-utils/dom';
+import createWrapper, { ElementWrapper, PropertyFilterWrapper } from '../../../lib/components/test-utils/dom';
 import PropertyFilter from '../../../lib/components/property-filter';
-
 import styles from '../../../lib/components/property-filter/styles.selectors.js';
 import { KeyCode } from '@cloudscape-design/test-utils-core/dist/utils';
 import {
@@ -20,7 +12,7 @@ import {
   PropertyFilterProps,
   Ref,
 } from '../../../lib/components/property-filter/interfaces';
-import TestI18nProvider from '../../../lib/components/i18n/testing';
+import { createDefaultProps } from './common';
 import '../../__a11y__/to-validate-a11y';
 
 const states: Record<string, string> = {
@@ -75,73 +67,21 @@ const filteringProperties: readonly FilteringProperty[] = [
 
 const filteringOptions: readonly FilteringOption[] = [
   { propertyKey: 'string', value: 'value1' },
-  { propertyKey: 'other-string', value: 'value1' },
   { propertyKey: 'string', value: 'value2' },
-  { propertyKey: 'range', value: '1' },
-  { propertyKey: 'range', value: '2' },
-  { propertyKey: 'other-string', value: 'value2' },
-  { propertyKey: 'missing-property', value: 'value' },
-  { propertyKey: 'default-operator', value: 'value' },
   { propertyKey: 'state', value: '0', label: getStateLabel('0') },
   { propertyKey: 'state', value: '1', label: getStateLabel('1') },
   { propertyKey: 'state', value: '2', label: getStateLabel('2') },
 ];
 
-const defaultProps: PropertyFilterProps = {
-  filteringEmpty: 'Empty',
-  filteringProperties,
-  filteringOptions,
-  filteringPlaceholder: 'Search',
-  filteringAriaLabel: 'your choice',
-  onChange: () => {},
-  query: { tokens: [], operation: 'and' },
-  i18nStrings,
-  id: 'property-filter',
-};
+const defaultProps = createDefaultProps(filteringProperties, filteringOptions);
 
 const renderComponent = (props?: Partial<PropertyFilterProps & { ref: React.Ref<Ref> }>) => {
-  const { container } = render(
-    <div>
-      <button id="button" />
-      <PropertyFilter {...defaultProps} {...props} />
-    </div>
-  );
-  const pageWrapper = createWrapper(container);
-  return { container, propertyFilterWrapper: pageWrapper.findPropertyFilter()!, pageWrapper };
+  const { container } = render(<PropertyFilter {...defaultProps} {...props} />);
+  return { container, propertyFilterWrapper: createWrapper(container).findPropertyFilter()! };
 };
 
-function StatefulPropertyFilter(props: Omit<PropertyFilterProps, 'onChange'>) {
-  const [query, setQuery] = useState<PropertyFilterProps.Query>(props.query);
-  return <PropertyFilter {...props} query={query} onChange={e => setQuery(e.detail)} />;
-}
-
-const renderStatefulComponent = (props?: Partial<PropertyFilterProps>) => {
-  const { container } = render(<StatefulPropertyFilter {...defaultProps} {...props} />);
-  const pageWrapper = createWrapper(container);
-  return { container, propertyFilterWrapper: pageWrapper.findPropertyFilter()!, pageWrapper };
-};
-
-function findPropertyField(wrapper: ElementWrapper) {
-  return wrapper.findFormField(`.${styles['token-editor-field-property']}`)!;
-}
-function findOperatorField(wrapper: ElementWrapper) {
-  return wrapper.findFormField(`.${styles['token-editor-field-operator']}`)!;
-}
 function findValueField(wrapper: ElementWrapper) {
   return wrapper.findFormField(`.${styles['token-editor-field-value']}`)!;
-}
-function findCancelButton(wrapper: ElementWrapper) {
-  return wrapper.findButton(`.${styles['token-editor-cancel']}`)!;
-}
-function findSubmitButton(wrapper: ElementWrapper) {
-  return wrapper.findButton(`.${styles['token-editor-submit']}`)!;
-}
-
-function findPropertySelector(wrapper: ElementWrapper) {
-  return findPropertyField(wrapper).findControl()!.findSelect()!;
-}
-function findOperatorSelector(wrapper: ElementWrapper) {
-  return findOperatorField(wrapper).findControl()!.findSelect()!;
 }
 function findValueSelector(wrapper: ElementWrapper) {
   return findValueField(wrapper).findControl()!.findAutosuggest()!;
@@ -156,752 +96,6 @@ function openTokenEditor(wrapper: PropertyFilterWrapper, index = 0) {
 }
 
 describe('property filter parts', () => {
-  describe('filtering input', () => {
-    describe('"use entered" option', () => {
-      test('is added, when a token can be created from the current text in the input', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent();
-        act(() => wrapper.setInputValue('string'));
-        expect(wrapper.findEnteredTextOption()!.getElement()).toHaveTextContent('Use: "string"');
-      });
-      test('with free text tokens disabled', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent({ disableFreeTextFiltering: true });
-        // is not shown, when the current text does not represent a property token
-        act(() => wrapper.setInputValue('string'));
-        expect(wrapper.findEnteredTextOption()).toBeNull();
-        // is shown otherwise
-        act(() => wrapper.setInputValue('string:'));
-        expect(wrapper.findEnteredTextOption()!.getElement()).toHaveTextContent('Use: "string:"');
-      });
-      test('causes a token to be created, when selected', () => {
-        const handleChange = jest.fn();
-        const { propertyFilterWrapper: wrapper } = renderComponent({ onChange: handleChange });
-        // free text
-        act(() => wrapper.setInputValue('string'));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.down));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.enter));
-        expect(handleChange).toHaveBeenCalledWith(
-          expect.objectContaining({ detail: { tokens: [{ value: 'string', operator: ':' }], operation: 'and' } })
-        );
-        // property
-        act(() => wrapper.setInputValue('string   !:123'));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.down));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.enter));
-        expect(handleChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            detail: { tokens: [{ propertyKey: 'string', value: '123', operator: '!:' }], operation: 'and' },
-          })
-        );
-      });
-    });
-    test('can be focused using the property filter ref', () => {
-      const focusRef: React.Ref<Ref> = { current: null };
-      const { propertyFilterWrapper: wrapper } = renderComponent({ ref: focusRef });
-      act(() => focusRef.current!.focus());
-      expect(wrapper.findNativeInput().getElement()).toHaveFocus();
-    });
-    test('recieves `placeholder`, `ariaLabel` and `disabled` properties passed to the component', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        disabled: true,
-        filteringPlaceholder: 'placeholder',
-      });
-      expect(wrapper.findNativeInput().getElement()).toHaveAttribute('placeholder', 'placeholder');
-      expect(wrapper.findNativeInput().getElement()).toHaveAttribute('disabled');
-      expect(wrapper.findNativeInput().getElement()).toHaveAttribute('aria-label', 'your choice');
-    });
-    test('recieves `ariaLabelledby`, `ariaDescribedby` and `controlId` properties passed to the component', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        ariaLabelledby: 'label-by-id',
-        ariaDescribedby: 'described-by-id',
-        controlId: 'control-id',
-      });
-      expect(wrapper.findNativeInput().getElement()).toHaveAttribute('aria-labelledby', 'label-by-id');
-      expect(wrapper.findNativeInput().getElement()).toHaveAttribute('aria-describedby', 'described-by-id');
-      expect(wrapper.findNativeInput().getElement()).toHaveAttribute('id', 'control-id');
-    });
-    describe('typing experience: ', () => {
-      test('provides relevant suggestions depending on the currently entered string', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent();
-        // property suggestions
-        act(() => wrapper.findNativeInput().focus());
-        expect(
-          wrapper
-            .findDropdown()
-            .findOptions()
-            .map(optionWrapper => optionWrapper.getElement().textContent)
-        ).toEqual(['string', 'string-other', 'default', 'string!=', 'state', 'range']);
-        // property and value suggestions
-        act(() => wrapper.setInputValue('a'));
-        expect(
-          wrapper
-            .findDropdown()
-            .findOptions()
-            .map(optionWrapper => optionWrapper.getElement().textContent)
-        ).toEqual([
-          'default',
-          'state',
-          'range',
-          'string = value1',
-          'string-other = value1',
-          'string = value2',
-          'string-other = value2',
-          'default = value',
-        ]);
-        // operator suggestions
-        act(() => wrapper.setInputValue('string'));
-        expect(
-          wrapper
-            .findDropdown()
-            .findOptions()
-            .map(optionWrapper => optionWrapper.getElement().textContent)
-        ).toEqual(['string =Equals', 'string !=Does not equal', 'string :Contains', 'string !:Does not contain']);
-        // value suggestions
-        act(() => wrapper.setInputValue('string:'));
-        expect(
-          wrapper
-            .findDropdown()
-            .findOptions()
-            .map(optionWrapper => optionWrapper.getElement().textContent)
-        ).toEqual(['string : value1', 'string : value2']);
-      });
-      test('supports property names that are substrings of others', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          filteringProperties: [
-            {
-              key: 'string',
-              propertyLabel: 'string',
-              operators: ['!:', ':', '=', '!='],
-              groupValuesLabel: 'String values',
-            },
-            {
-              key: 'other-string',
-              propertyLabel: 'string-other',
-              operators: ['!:', ':', '=', '!='],
-              groupValuesLabel: 'Other string values',
-            },
-          ],
-        });
-        act(() => wrapper.setInputValue('string'));
-        expect(
-          wrapper
-            .findDropdown()
-            .findOptions()
-            .map(optionWrapper => optionWrapper.getElement().textContent)
-        ).toEqual(['string =Equals', 'string !=Does not equal', 'string :Contains', 'string !:Does not contain']);
-        act(() => wrapper.setInputValue('string-other'));
-        expect(
-          wrapper
-            .findDropdown()
-            .findOptions()
-            .map(optionWrapper => optionWrapper.getElement().textContent)
-        ).toEqual([
-          'string-other =Equals',
-          'string-other !=Does not equal',
-          'string-other :Contains',
-          'string-other !:Does not contain',
-        ]);
-      });
-      test('supports property names that are substrings of others (alternate order)', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          filteringProperties: [
-            {
-              key: 'other-string',
-              propertyLabel: 'string-other',
-              operators: ['!:', ':', '=', '!='],
-              groupValuesLabel: 'Other string values',
-            },
-            {
-              key: 'string',
-              propertyLabel: 'string',
-              operators: ['!:', ':', '=', '!='],
-              groupValuesLabel: 'String values',
-            },
-          ],
-        });
-        act(() => wrapper.setInputValue('string'));
-        expect(
-          wrapper
-            .findDropdown()
-            .findOptions()
-            .map(optionWrapper => optionWrapper.getElement().textContent)
-        ).toEqual(['string =Equals', 'string !=Does not equal', 'string :Contains', 'string !:Does not contain']);
-        act(() => wrapper.setInputValue('string-other'));
-        expect(
-          wrapper
-            .findDropdown()
-            .findOptions()
-            .map(optionWrapper => optionWrapper.getElement().textContent)
-        ).toEqual([
-          'string-other =Equals',
-          'string-other !=Does not equal',
-          'string-other :Contains',
-          'string-other !:Does not contain',
-        ]);
-      });
-      test('can be submitted to create a token', () => {
-        const handleChange = jest.fn();
-        const { propertyFilterWrapper: wrapper } = renderComponent({ onChange: handleChange });
-        act(() => wrapper.findNativeInput().focus());
-        act(() => wrapper.setInputValue('string'));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.enter));
-        expect(handleChange).toHaveBeenCalledWith(
-          expect.objectContaining({
-            detail: { tokens: [{ value: 'string', operator: ':' }], operation: 'and' },
-          })
-        );
-        expect(wrapper.findNativeInput().getElement()).toHaveValue('');
-
-        act(() => wrapper.setInputValue('string:value'));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.enter));
-        expect(handleChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            detail: { tokens: [{ value: 'value', operator: ':', propertyKey: 'string' }], operation: 'and' },
-          })
-        );
-        expect(wrapper.findNativeInput().getElement()).toHaveValue('');
-      });
-      test('submitting empty input does not create a token', () => {
-        const handleChange = jest.fn();
-        const { propertyFilterWrapper: wrapper } = renderComponent({ onChange: handleChange });
-        act(() => wrapper.findNativeInput().focus());
-        act(() => wrapper.findNativeInput().keydown(KeyCode.enter));
-        expect(handleChange).not.toHaveBeenCalled();
-      });
-      test('can not be submitted to create a free text token, when free text tokens are disabled', () => {
-        const handleChange = jest.fn();
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          onChange: handleChange,
-          disableFreeTextFiltering: true,
-        });
-        act(() => wrapper.findNativeInput().focus());
-        act(() => wrapper.setInputValue('string'));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.enter));
-        expect(handleChange).not.toHaveBeenCalled();
-        expect(wrapper.findNativeInput().getElement()).toHaveValue('string');
-      });
-    });
-    describe('selecting items in the dropdown', () => {
-      test('keeps the dropdown open, when property or operator suggestions are selected', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent();
-
-        act(() => wrapper.findNativeInput().focus());
-        act(() => wrapper.findNativeInput().keydown(KeyCode.down));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.enter));
-        expect(wrapper.findNativeInput().getElement()).toHaveValue('string');
-        expect(wrapper.findDropdown()?.findOpenDropdown()).toBeTruthy();
-
-        act(() => wrapper.findNativeInput().keydown(KeyCode.down));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.down));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.enter));
-        expect(wrapper.findNativeInput().getElement()).toHaveValue('string = ');
-        expect(wrapper.findDropdown()?.findOpenDropdown()).toBeTruthy();
-      });
-      test('creates a new token and closes the dropdown, when a value suggestion is selected', () => {
-        const handleChange = jest.fn();
-        const { propertyFilterWrapper: wrapper } = renderComponent({ onChange: handleChange });
-
-        act(() => wrapper.setInputValue('string='));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.down));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.down));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.enter));
-        expect(wrapper.findNativeInput().getElement()).toHaveValue('');
-        expect(handleChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            detail: { tokens: [{ value: 'value1', operator: '=', propertyKey: 'string' }], operation: 'and' },
-          })
-        );
-        expect(wrapper.findDropdown()?.findOpenDropdown()).toBeFalsy();
-      });
-      test('inserts operator when property is selected if only one operator is defined for that property', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent();
-
-        act(() => wrapper.findNativeInput().focus());
-        act(() => wrapper.findNativeInput().keydown(KeyCode.down));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.down));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.down));
-        act(() => wrapper.findNativeInput().keydown(KeyCode.enter));
-        expect(wrapper.findNativeInput().getElement()).toHaveValue('default = ');
-        expect(wrapper.findDropdown()?.findOpenDropdown()).toBeTruthy();
-      });
-    });
-
-    test('can use selectSuggestionByValue', () => {
-      const onChange = jest.fn();
-      const { propertyFilterWrapper: wrapper } = renderComponent({ onChange });
-      act(() => wrapper.focus());
-      act(() => wrapper.selectSuggestionByValue('string-other'));
-      expect(wrapper.findEnteredTextOption()!.getElement()).toHaveTextContent('Use: "string-other"');
-
-      act(() => wrapper.selectSuggestionByValue('string-other != '));
-      expect(wrapper.findEnteredTextOption()!.getElement()).toHaveTextContent('Use: "string-other != "');
-
-      act(() => wrapper.selectSuggestionByValue('string-other != value2'));
-      expect(onChange).toHaveBeenCalledWith(
-        expect.objectContaining({
-          detail: { operation: 'and', tokens: [{ propertyKey: 'other-string', operator: '!=', value: 'value2' }] },
-        })
-      );
-    });
-  });
-
-  describe('count text', () => {
-    test('is not displayed when count text is empty', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        countText: '',
-        query: { tokens: [{ propertyKey: 'string', value: 'first', operator: ':' }], operation: 'or' },
-      });
-      expect(wrapper.findResultsCount()).toBe(null);
-    });
-    test('is not displayed when there are no filtering tokens', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({ countText: '5 matches' });
-      expect(wrapper.findResultsCount()).toBe(null);
-    });
-    test('is visible when there is at least 1 token', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        countText: '5 matches',
-        query: { tokens: [{ propertyKey: 'string', value: 'first', operator: ':' }], operation: 'or' },
-      });
-      expect(wrapper.findResultsCount()!.getElement()).toHaveTextContent('5 matches');
-    });
-  });
-
-  describe('constraint text', () => {
-    test('is not displayed when constraint text is empty', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        filteringConstraintText: '',
-        query: { tokens: [{ propertyKey: 'string', value: 'first', operator: ':' }], operation: 'or' },
-      });
-      expect(wrapper.findConstraint()).toBe(null);
-    });
-    test('is visible when the constraint text is not empty', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        filteringConstraintText: <div>This is my constraint</div>,
-        query: { tokens: [{ propertyKey: 'string', value: 'first', operator: ':' }], operation: 'or' },
-      });
-      expect(wrapper.findConstraint()!.getElement()).toHaveTextContent('This is my constraint');
-    });
-    test('is used as ARIA description for the autosuggest input', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        ariaDescribedby: 'my-custom-description',
-        filteringConstraintText: <div>This is my constraint</div>,
-        query: { tokens: [{ propertyKey: 'string', value: 'first', operator: ':' }], operation: 'or' },
-      });
-      expect(wrapper.findNativeInput().getElement()).toHaveAttribute(
-        'aria-describedby',
-        'my-custom-description' + ' ' + wrapper.findConstraint()!.getElement().id
-      );
-    });
-  });
-
-  describe('filtering  tokens', () => {
-    describe('content', () => {
-      test('free text token', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          query: { tokens: [{ value: 'first', operator: ':' }], operation: 'or' },
-        });
-        expect(wrapper.findTokens()![0].getElement()).toHaveTextContent('first');
-      });
-      test('negated free text token', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          query: { tokens: [{ value: 'first', operator: '!:' }], operation: 'or' },
-        });
-        expect(wrapper.findTokens()![0].getElement()).toHaveTextContent('!: first');
-      });
-      test('custom operator free text token', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          freeTextFiltering: { operators: ['!^'] },
-          query: { tokens: [{ value: 'first', operator: '!^' }], operation: 'or' },
-        });
-        expect(wrapper.findTokens()![0].getElement()).toHaveTextContent('!^ first');
-      });
-      test('property token', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          query: { tokens: [{ propertyKey: 'range', value: 'value', operator: '>' }], operation: 'or' },
-        });
-        expect(wrapper.findTokens()![0].getElement()).toHaveTextContent('range > value');
-      });
-    });
-    describe('join operation control', () => {
-      let wrapper: PropertyFilterWrapper;
-      const handleChange = jest.fn();
-      beforeEach(() => {
-        wrapper = renderComponent({
-          query: {
-            tokens: [
-              { propertyKey: 'range', value: 'value', operator: '>' },
-              { propertyKey: 'string', value: 'value1', operator: ':' },
-            ],
-            operation: 'or',
-          },
-          onChange: handleChange,
-        }).propertyFilterWrapper;
-        handleChange.mockReset();
-      });
-      test('is missing on the first token', () => {
-        expect(wrapper.findTokens()![0].findTokenOperation()).toBeNull();
-      });
-      test('can be used to change the operation', () => {
-        const secondToken = wrapper.findTokens()![1];
-        act(() => secondToken.findTokenOperation()!.openDropdown());
-        act(() => secondToken.findTokenOperation()!.selectOption(1));
-        expect(handleChange).toHaveBeenCalledWith(
-          expect.objectContaining({ detail: expect.objectContaining({ operation: 'and' }) })
-        );
-      });
-
-      test('can be used to check aria-label attirbute on the filter token operator', () => {
-        const secondToken = wrapper.findTokens()![1];
-        const label = secondToken
-          .findTokenOperation()
-          ?.findTrigger()
-          .getElement()
-          .getAttribute('aria-labelledby')!
-          .split(' ')
-          .map(labelId => wrapper.getElement().querySelector(`#${labelId}`)!.textContent)
-          .join(' ');
-        expect(label).toBe('Boolean Operator or');
-      });
-    });
-    describe('dismiss button', () => {
-      test('causes onChange to fire, removing the token', () => {
-        const handleChange = jest.fn();
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          onChange: handleChange,
-          query: { tokens: [{ value: 'first', operator: '!:' }], operation: 'or' },
-        });
-        act(() => wrapper.findTokens()![0].findRemoveButton()!.click());
-        expect(handleChange).toHaveBeenCalledWith(expect.objectContaining({ detail: { tokens: [], operation: 'or' } }));
-      });
-      test('moves the focus to the input, when pressed', () => {
-        const { propertyFilterWrapper: wrapper } = renderStatefulComponent({
-          query: { tokens: [{ value: 'first', operator: '!:' }], operation: 'or' },
-        });
-        act(() => wrapper.findTokens()![0].findRemoveButton()!.click());
-        expect(wrapper.findNativeInput().getElement()).toHaveFocus();
-      });
-      test('has a label from i18nStrings', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          query: { tokens: [{ value: 'first', operator: '!:' }], operation: 'or' },
-        });
-        expect(wrapper.findTokens()![0].findRemoveButton().getElement()).toHaveAttribute(
-          'aria-label',
-          'Remove token undefined !: first'
-        );
-      });
-      test('disabled, when the component is disabled', () => {
-        const handleChange = jest.fn();
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          onChange: handleChange,
-          query: { tokens: [{ value: 'first', operator: '!:' }], operation: 'or' },
-          disabled: true,
-        });
-        expect(wrapper.findTokens()![0].findRemoveButton().getElement()).toBeDisabled();
-        act(() => wrapper.findTokens()![0].findRemoveButton()!.click());
-        expect(handleChange).not.toHaveBeenCalled();
-      });
-    });
-  });
-
-  describe('token editor', () => {
-    test('uses i18nStrings to populate header', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        query: { tokens: [{ value: 'first', operator: '!:' }], operation: 'or' },
-      });
-      const [, popoverWrapper] = openTokenEditor(wrapper);
-      expect(popoverWrapper.findHeader()!.getElement()).toHaveTextContent(i18nStrings.editTokenHeader);
-    });
-    describe('form controls content', () => {
-      test('default', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          query: { tokens: [{ propertyKey: 'string', value: 'first', operator: '!:' }], operation: 'or' },
-        });
-        const [contentWrapper] = openTokenEditor(wrapper);
-        expect(contentWrapper.getElement()).toHaveTextContent(
-          'PropertystringOperator!:Does not containValueCancelApply'
-        );
-        const propertySelectWrapper = findPropertySelector(contentWrapper);
-        act(() => propertySelectWrapper.openDropdown());
-        expect(
-          propertySelectWrapper
-            .findDropdown()
-            .findOptions()!
-            .map(optionWrapper => optionWrapper.getElement().textContent)
-        ).toEqual(['All properties', 'string', 'string-other', 'default', 'string!=', 'state', 'range']);
-
-        const operatorSelectWrapper = findOperatorSelector(contentWrapper);
-        act(() => operatorSelectWrapper.openDropdown());
-        expect(
-          operatorSelectWrapper
-            .findDropdown()
-            .findOptions()!
-            .map(optionWrapper => optionWrapper.getElement().textContent)
-        ).toEqual(['=Equals', '!=Does not equal', ':Contains', '!:Does not contain']);
-
-        const valueSelectWrapper = findValueSelector(contentWrapper);
-        act(() => valueSelectWrapper.focus());
-        expect(
-          valueSelectWrapper
-            .findDropdown()
-            .findOptions()!
-            .map(optionWrapper => optionWrapper.getElement().textContent)
-        ).toEqual(['value1', 'value2']);
-      });
-      test('with free text disabled', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          disableFreeTextFiltering: true,
-          query: { tokens: [{ propertyKey: 'string', value: 'first', operator: '!:' }], operation: 'or' },
-        });
-        const [contentWrapper] = openTokenEditor(wrapper);
-        expect(contentWrapper.getElement()).toHaveTextContent(
-          'PropertystringOperator!:Does not containValueCancelApply'
-        );
-        const propertySelectWrapper = findPropertySelector(contentWrapper);
-        act(() => propertySelectWrapper.openDropdown());
-        expect(
-          propertySelectWrapper
-            .findDropdown()
-            .findOptions()!
-            .map(optionWrapper => optionWrapper.getElement().textContent)
-        ).toEqual(['string', 'string-other', 'default', 'string!=', 'state', 'range']);
-      });
-      test('with custom free text operators', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          freeTextFiltering: { operators: ['=', '!=', ':', '!:'] },
-          query: { tokens: [{ value: 'first', operator: '!=' }], operation: 'or' },
-        });
-        const [contentWrapper] = openTokenEditor(wrapper);
-        expect(contentWrapper.getElement()).toHaveTextContent(
-          'PropertyAll propertiesOperator!=Does not equalValueCancelApply'
-        );
-        const operatorSelectWrapper = findOperatorSelector(contentWrapper);
-        act(() => operatorSelectWrapper.openDropdown());
-        expect(
-          operatorSelectWrapper
-            .findDropdown()
-            .findOptions()!
-            .map(optionWrapper => optionWrapper.getElement().textContent)
-        ).toEqual(['=Equals', '!=Does not equal', ':Contains', '!:Does not contain']);
-      });
-      test('preserves fields, when one is edited', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          disableFreeTextFiltering: true,
-          query: { tokens: [{ propertyKey: 'string', value: 'first', operator: ':' }], operation: 'or' },
-        });
-        const [contentWrapper] = openTokenEditor(wrapper);
-        const propertySelectWrapper = findPropertySelector(contentWrapper);
-        const operatorSelectWrapper = findOperatorSelector(contentWrapper);
-        const valueSelectWrapper = findValueSelector(contentWrapper);
-
-        // Change operator
-        act(() => operatorSelectWrapper.openDropdown());
-        act(() => operatorSelectWrapper.selectOption(1));
-        expect(propertySelectWrapper.findTrigger().getElement()).toHaveTextContent('string');
-        expect(operatorSelectWrapper.findTrigger().getElement()).toHaveTextContent('=Equals');
-        expect(valueSelectWrapper.findNativeInput().getElement()).toHaveAttribute('value', 'first');
-
-        // Change value
-        act(() => valueSelectWrapper.setInputValue('123'));
-        expect(propertySelectWrapper.findTrigger().getElement()).toHaveTextContent('string');
-        expect(operatorSelectWrapper.findTrigger().getElement()).toHaveTextContent('=Equals');
-        expect(valueSelectWrapper.findNativeInput().getElement()).toHaveAttribute('value', '123');
-
-        // Change property
-        // This preserves operator but not value because the value type between properties can be different.
-        act(() => propertySelectWrapper.openDropdown());
-        act(() => propertySelectWrapper.selectOption(2));
-        expect(propertySelectWrapper.findTrigger().getElement()).toHaveTextContent('string-other');
-        expect(operatorSelectWrapper.findTrigger().getElement()).toHaveTextContent('=Equals');
-        expect(valueSelectWrapper.findNativeInput().getElement()).toHaveAttribute('value', '');
-      });
-    });
-    describe('exit actions', () => {
-      let wrapper: PropertyFilterWrapper;
-      let contentWrapper: ElementWrapper;
-      let popoverWrapper: PopoverWrapper;
-      const handleChange = jest.fn();
-      beforeEach(() => {
-        handleChange.mockReset();
-        const { propertyFilterWrapper } = renderComponent({
-          onChange: handleChange,
-          query: { tokens: [{ propertyKey: 'string', value: 'first', operator: '!:' }], operation: 'or' },
-        });
-        wrapper = propertyFilterWrapper;
-        const openResult = openTokenEditor(wrapper);
-        contentWrapper = openResult[0];
-        popoverWrapper = openResult[1];
-      });
-      test('dismiss button closes the popover and discards the changes', () => {
-        const valueAutosuggestWrapper = findValueSelector(contentWrapper);
-        act(() => valueAutosuggestWrapper.setInputValue('123'));
-        const operatorSelectWrapper = findOperatorSelector(contentWrapper);
-        act(() => operatorSelectWrapper.openDropdown());
-        act(() => operatorSelectWrapper.selectOption(1));
-        const propertySelectWrapper = findPropertySelector(contentWrapper);
-        act(() => propertySelectWrapper.openDropdown());
-        act(() => propertySelectWrapper.selectOption(1));
-
-        act(() => popoverWrapper.findDismissButton()!.click());
-        expect(popoverWrapper.findContent()).toBeNull();
-        expect(handleChange).not.toHaveBeenCalled();
-
-        const openResult = openTokenEditor(wrapper);
-        contentWrapper = openResult[0];
-        expect(contentWrapper.getElement()).toHaveTextContent(
-          'PropertystringOperator!:Does not containValueCancelApply'
-        );
-      });
-      test('cancel button closes the popover and discards the changes', () => {
-        const valueAutosuggestWrapper = findValueSelector(contentWrapper);
-        act(() => valueAutosuggestWrapper.setInputValue('123'));
-        const operatorSelectWrapper = findOperatorSelector(contentWrapper);
-        act(() => operatorSelectWrapper.openDropdown());
-        act(() => operatorSelectWrapper.selectOption(1));
-        const propertySelectWrapper = findPropertySelector(contentWrapper);
-        act(() => propertySelectWrapper.openDropdown());
-        act(() => propertySelectWrapper.selectOption(1));
-
-        act(() => contentWrapper.findButton(`.${styles['token-editor-cancel']}`)!.click());
-        popoverWrapper = wrapper.findTokens()[0]!.find('*')!.findPopover()!;
-        expect(popoverWrapper.findContent()).toBeNull();
-        expect(handleChange).not.toHaveBeenCalled();
-
-        const openResult = openTokenEditor(wrapper);
-        contentWrapper = openResult[0];
-        expect(contentWrapper.getElement()).toHaveTextContent(
-          'PropertystringOperator!:Does not containValueCancelApply'
-        );
-      });
-      test('submit button closes the popover and saves the changes', () => {
-        const operatorSelectWrapper = findOperatorSelector(contentWrapper);
-        act(() => operatorSelectWrapper.openDropdown());
-        act(() => operatorSelectWrapper.selectOption(1));
-        const propertySelectWrapper = findPropertySelector(contentWrapper);
-        act(() => propertySelectWrapper.openDropdown());
-        act(() => propertySelectWrapper.selectOption(1));
-        const valueAutosuggestWrapper = findValueSelector(contentWrapper);
-        act(() => valueAutosuggestWrapper.setInputValue('123'));
-
-        act(() => contentWrapper.findButton(`.${styles['token-editor-submit']}`)!.click());
-        popoverWrapper = wrapper.findTokens()[0]!.find('*')!.findPopover()!;
-        expect(popoverWrapper.findContent()).toBeNull();
-        expect(handleChange).toHaveBeenCalledWith(
-          expect.objectContaining({
-            detail: { operation: 'or', tokens: [{ operator: ':', propertyKey: undefined, value: '123' }] },
-          })
-        );
-      });
-    });
-  });
-  describe('tokens collapsed/expanded toggle', () => {
-    test('is hidden, when there are no filtering tokens', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({ tokenLimit: 0 });
-      expect(wrapper.findTokenToggle()).toBeNull();
-    });
-    test('is hidden, when tokenLimit is not provided', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        query: { tokens: [{ propertyKey: 'string', value: 'first', operator: ':' }], operation: 'or' },
-      });
-      expect(wrapper.findTokenToggle()).toBeNull();
-    });
-
-    describe('aria-label on toggle fewer/more button', () => {
-      const propertiesWithMultipleTokens: Partial<PropertyFilterProps> = {
-        tokenLimit: 1,
-        query: {
-          tokens: [
-            { propertyKey: 'string', value: 'first', operator: ':' },
-            { propertyKey: 'string', value: 'second', operator: ':' },
-          ],
-          operation: 'or',
-        },
-      };
-
-      test('sets no aria-label text on the expand button by default', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent(propertiesWithMultipleTokens);
-        expect(wrapper.findTokenToggle()!.getElement()).not.toHaveAttribute('aria-label');
-      });
-      test('sets aria-label text on the expand button when provided', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          ...propertiesWithMultipleTokens,
-          tokenLimitShowMoreAriaLabel: 'Show more token',
-        });
-        expect(wrapper.findTokenToggle()!.getElement()).toHaveAttribute('aria-label', 'Show more token');
-      });
-      test('sets no aria-label text on the collapse button by default', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent(propertiesWithMultipleTokens);
-        wrapper.findTokenToggle()!.click();
-        expect(wrapper.findTokenToggle()!.getElement()).not.toHaveAttribute('aria-label');
-      });
-      test('sets aria-label text on the collapse button when provided', () => {
-        const { propertyFilterWrapper: wrapper } = renderComponent({
-          ...propertiesWithMultipleTokens,
-          tokenLimitShowFewerAriaLabel: 'Show fewer token',
-        });
-        wrapper.findTokenToggle()!.click();
-        expect(wrapper.findTokenToggle()!.getElement()).toHaveAttribute('aria-label', 'Show fewer token');
-      });
-    });
-
-    test('toggles the visibility of tokens past the limit', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        tokenLimit: 1,
-        query: {
-          tokens: [
-            { propertyKey: 'string', value: 'first', operator: ':' },
-            { propertyKey: 'string', value: 'second', operator: ':' },
-          ],
-          operation: 'or',
-        },
-      });
-      expect(wrapper.findTokenToggle()!.getElement()).toHaveTextContent(i18nStrings.tokenLimitShowMore);
-      expect(wrapper.findTokens()!).toHaveLength(1);
-      expect(wrapper.findTokens()![0].findLabel().getElement()).toHaveTextContent('string : first');
-      // show more
-      act(() => wrapper.findTokenToggle()!.click());
-      expect(wrapper.findTokenToggle()!.getElement()).toHaveTextContent(i18nStrings.tokenLimitShowFewer);
-      expect(wrapper.findTokens()!).toHaveLength(2);
-      // show fewer
-      act(() => wrapper.findTokenToggle()!.click());
-      expect(wrapper.findTokenToggle()!.getElement()).toHaveTextContent(i18nStrings.tokenLimitShowMore);
-      expect(wrapper.findTokens()!).toHaveLength(1);
-    });
-  });
-
-  describe('"remove all tokens" button', () => {
-    test('is hidden, when there are no filtering tokens', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent();
-      expect(wrapper.findRemoveAllButton()).toBeNull();
-    });
-    test('a11y label', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        query: { tokens: [{ propertyKey: 'string', value: 'first', operator: ':' }], operation: 'or' },
-      });
-      expect(wrapper.findRemoveAllButton()!.getElement()).toHaveTextContent(i18nStrings.clearFiltersText);
-    });
-    test('causes onChange to fire, removing all tokens', () => {
-      const spy = jest.fn();
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        query: { tokens: [{ propertyKey: 'string', value: 'first', operator: ':' }], operation: 'or' },
-        onChange: spy,
-      });
-      act(() => wrapper.findRemoveAllButton()!.click());
-      expect(spy).toHaveBeenCalledWith(expect.objectContaining({ detail: { tokens: [], operation: 'or' } }));
-    });
-    test('moves focus to the input but keeps dropdown closed', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        query: { tokens: [{ propertyKey: 'string', value: 'first', operator: ':' }], operation: 'or' },
-      });
-      act(() => wrapper.findRemoveAllButton()!.click());
-      expect(wrapper.findNativeInput()!.getElement()).toHaveFocus();
-      expect(wrapper.findDropdown().findOpenDropdown()).toBe(null);
-    });
-    test('disabled, when the component is disabled', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        disabled: true,
-        query: { tokens: [{ propertyKey: 'string', value: 'first', operator: ':' }], operation: 'or' },
-      });
-      expect(wrapper.findRemoveAllButton()!.getElement()).toBeDisabled();
-    });
-  });
   test('`expandToViewport` property test utils', () => {
     const { propertyFilterWrapper: wrapper } = renderComponent({
       expandToViewport: true,
@@ -959,139 +153,6 @@ describe('property filter parts', () => {
       wrapper.findNativeInput().focus();
       wrapper.setInputValue('string');
       expect(wrapper.findDropdown().find('ul')!.getElement()).toHaveAccessibleDescription(`Finished text`);
-    });
-  });
-
-  describe('extended operators', () => {
-    const indexProperty: FilteringProperty = {
-      key: 'index',
-      operators: [
-        {
-          operator: '>',
-          form: ({ value, onChange, operator, filter = '' }) => (
-            <button data-testid="change+" data-context={operator + filter} onClick={() => onChange((value || 0) + 1)}>
-              {value || 0}
-            </button>
-          ),
-        },
-        {
-          operator: '<',
-          form: ({ value, onChange, operator, filter = '' }) => (
-            <button data-testid="change-" data-context={operator + filter} onClick={() => onChange((value || 0) - 1)}>
-              {value || 0}
-            </button>
-          ),
-        },
-      ],
-      propertyLabel: 'index',
-      groupValuesLabel: 'index value',
-    };
-    const extendedOperatorProps = { filteringProperties: [indexProperty] };
-
-    test('property filter renders operator form instead of options list', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent(extendedOperatorProps);
-      wrapper.setInputValue('index >');
-      expect(wrapper.findDropdown()!.findOpenDropdown()!.find('[data-testid="change+"]')).not.toBe(null);
-      wrapper.setInputValue('index <');
-      expect(wrapper.findDropdown()!.findOpenDropdown()!.find('[data-testid="change-"]')).not.toBe(null);
-    });
-
-    test('property filter uses operator form in the token editor', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        ...extendedOperatorProps,
-        query: {
-          tokens: [
-            { propertyKey: 'index', value: 1, operator: '>' },
-            { propertyKey: 'index', value: 2, operator: '<' },
-          ],
-          operation: 'and',
-        },
-      });
-      expect(wrapper.findTokens()).toHaveLength(2);
-      expect(openTokenEditor(wrapper, 0)[0].find('[data-testid="change+"]')).not.toBe(null);
-      expect(openTokenEditor(wrapper, 1)[0].find('[data-testid="change-"]')).not.toBe(null);
-    });
-
-    test('extended operator form takes value/onChange state', () => {
-      const { propertyFilterWrapper: wrapper, pageWrapper } = renderComponent(extendedOperatorProps);
-      wrapper.setInputValue('index >');
-      expect(pageWrapper.find('[data-testid="change+"]')!.getElement()).toHaveTextContent('0');
-      act(() => pageWrapper.find('[data-testid="change+"]')!.click());
-      expect(pageWrapper.find('[data-testid="change+"]')!.getElement()).toHaveTextContent('1');
-    });
-
-    test('extended operator form can be cancelled/submitted', () => {
-      const onChange = jest.fn();
-      const { propertyFilterWrapper: wrapper, pageWrapper } = renderComponent({ ...extendedOperatorProps, onChange });
-
-      // Increment value
-      wrapper.setInputValue('index >');
-      act(() => pageWrapper.find('[data-testid="change+"]')!.click());
-
-      // Click cancel
-      act(() => pageWrapper.findButton(`.${styles['property-editor-cancel']}`)!.click());
-      expect(wrapper.findDropdown()!.findOpenDropdown()).toBe(null);
-      expect(onChange).not.toBeCalled();
-      expect(wrapper.findNativeInput().getElement()).toHaveFocus();
-
-      // Decrement value
-      wrapper.setInputValue('index <');
-      act(() => pageWrapper.find('[data-testid="change-"]')!.click());
-
-      // Click submit
-      act(() => pageWrapper.findButton(`.${styles['property-editor-submit']}`)!.click());
-      expect(wrapper.findDropdown()!.findOpenDropdown()).toBe(null);
-      expect(onChange).toBeCalledWith(
-        expect.objectContaining({
-          detail: {
-            operation: 'and',
-            tokens: [{ propertyKey: 'index', operator: '<', value: -1 }],
-          },
-        })
-      );
-      expect(wrapper.findNativeInput().getElement()).toHaveFocus();
-    });
-
-    test('extended operator form takes chosen operator and entered filter text', () => {
-      const { propertyFilterWrapper: wrapper, pageWrapper } = renderComponent(extendedOperatorProps);
-
-      wrapper.setInputValue('index >');
-      expect(pageWrapper.find('[data-testid="change+"]')!.getElement()).toHaveAttribute('data-context', '>');
-
-      wrapper.setInputValue('index <');
-      expect(pageWrapper.find('[data-testid="change-"]')!.getElement()).toHaveAttribute('data-context', '<');
-
-      wrapper.setInputValue('index < ');
-      expect(pageWrapper.find('[data-testid="change-"]')!.getElement()).toHaveAttribute('data-context', '<');
-
-      wrapper.setInputValue('index < 1 2 ');
-      expect(pageWrapper.find('[data-testid="change-"]')!.getElement()).toHaveAttribute('data-context', '<1 2 ');
-    });
-
-    test('property filter uses operator formatter to render token value', () => {
-      const { propertyFilterWrapper: wrapper } = renderComponent({
-        filteringProperties: [
-          {
-            key: 'index',
-            operators: [
-              { operator: '=', format: value => `EQ${value}` },
-              { operator: '!=', format: value => `NEQ${value}` },
-            ],
-            propertyLabel: 'index',
-            groupValuesLabel: 'index value',
-          },
-        ],
-        query: {
-          tokens: [
-            { propertyKey: 'index', value: 1, operator: '=' },
-            { propertyKey: 'index', value: 2, operator: '!=' },
-          ],
-          operation: 'and',
-        },
-      });
-      expect(wrapper.findTokens()).toHaveLength(2);
-      expect(wrapper.findTokens()[0].getElement()).toHaveTextContent('index = EQ1');
-      expect(wrapper.findTokens()[1].getElement()).toHaveTextContent('index != NEQ2');
     });
   });
 
@@ -1379,182 +440,5 @@ describe('property filter parts', () => {
   test('check a11y', async () => {
     const { container } = render(<PropertyFilter {...defaultProps} />);
     await expect(container).toValidateA11y();
-  });
-});
-
-describe('i18n', () => {
-  const providerMessages = {
-    'property-filter': {
-      'i18nStrings.allPropertiesLabel': 'Custom All properties',
-      'i18nStrings.groupPropertiesText': 'Custom Properties',
-      'i18nStrings.groupValuesText': 'Custom Values',
-      'i18nStrings.operatorContainsText': 'Custom Contains',
-      'i18nStrings.operatorDoesNotContainText': 'Custom Does not contain',
-      'i18nStrings.operatorDoesNotEqualText': 'Custom Does not equal',
-      'i18nStrings.operatorEqualsText': 'Custom Equals',
-      'i18nStrings.operatorGreaterOrEqualText': 'Custom Greater than or equal',
-      'i18nStrings.operatorGreaterText': 'Custom Greater than',
-      'i18nStrings.operatorLessOrEqualText': 'Custom Less than or equal',
-      'i18nStrings.operatorLessText': 'Custom Less than',
-      'i18nStrings.operatorStartsWithText': 'Custom Starts with',
-      'i18nStrings.operatorDoesNotStartWithText': 'Custom Does not start with',
-      'i18nStrings.operatorText': 'Custom Operator',
-      'i18nStrings.operatorsText': 'Custom Operators',
-      'i18nStrings.propertyText': 'Custom Property',
-    },
-  };
-
-  it('uses dropdown labels from i18n provider for a string property', () => {
-    const { container } = render(
-      <TestI18nProvider messages={providerMessages}>
-        <PropertyFilter
-          {...defaultProps}
-          filteringProperties={[
-            {
-              key: 'string',
-              propertyLabel: 'string',
-              operators: ['!:', ':', '=', '!=', '^', '!^'],
-              groupValuesLabel: 'String values',
-            },
-          ]}
-          i18nStrings={{ filteringAriaLabel: 'your choice', filteringPlaceholder: 'Search' }}
-        />
-      </TestI18nProvider>
-    );
-    const wrapper = createWrapper(container).findPropertyFilter()!;
-
-    wrapper.focus();
-    const dropdown = wrapper.findDropdown()!;
-    expect(dropdown.find('li')!.getElement()).toHaveTextContent('Custom Properties');
-
-    wrapper.selectSuggestion(1);
-    expect(dropdown.find('li:nth-child(2)')!.getElement()).toHaveTextContent('Custom Operators');
-    expect(
-      dropdown.findOptions().map(optionWrapper => optionWrapper.findDescription()?.getElement().textContent)
-    ).toEqual([
-      'Custom Equals',
-      'Custom Does not equal',
-      'Custom Contains',
-      'Custom Does not contain',
-      'Custom Starts with',
-      'Custom Does not start with',
-    ]);
-  });
-
-  it('uses dropdown labels from i18n provider for a numeric property', () => {
-    const { container } = render(
-      <TestI18nProvider messages={providerMessages}>
-        <PropertyFilter
-          {...defaultProps}
-          i18nStrings={{ filteringAriaLabel: 'your choice', filteringPlaceholder: 'Search' }}
-        />
-      </TestI18nProvider>
-    );
-    const wrapper = createWrapper(container).findPropertyFilter()!;
-
-    wrapper.focus();
-    const dropdown = wrapper.findDropdown()!;
-    expect(dropdown.find('li')!.getElement()).toHaveTextContent('Custom Properties');
-
-    wrapper.selectSuggestion(6);
-    expect(dropdown.find('li:nth-child(2)')!.getElement()).toHaveTextContent('Custom Operators');
-    expect(
-      dropdown.findOptions().map(optionWrapper => optionWrapper.findDescription()?.getElement().textContent)
-    ).toEqual([
-      'Custom Equals',
-      'Custom Does not equal',
-      'Custom Greater than or equal',
-      'Custom Less than or equal',
-      'Custom Less than',
-      'Custom Greater than',
-    ]);
-  });
-
-  it('uses token and editor labels from i18n provider', () => {
-    const { container } = render(
-      <TestI18nProvider
-        messages={{
-          'property-filter': {
-            'i18nStrings.applyActionText': 'Custom Apply',
-            'i18nStrings.cancelActionText': 'Custom Cancel',
-            'i18nStrings.clearFiltersText': 'Custom Clear filters',
-            'i18nStrings.editTokenHeader': 'Custom Edit filter',
-            'i18nStrings.groupPropertiesText': 'Custom Properties',
-            'i18nStrings.groupValuesText': 'Custom Values',
-            'i18nStrings.operationAndText': 'Custom and',
-            'i18nStrings.operationOrText': 'Custom or',
-            'i18nStrings.operatorText': 'Custom Operator',
-            'i18nStrings.operatorsText': 'Operators',
-            'i18nStrings.propertyText': 'Custom Property',
-            'i18nStrings.tokenLimitShowFewer': 'Custom Show fewer',
-            'i18nStrings.tokenLimitShowMore': 'Custom Show more',
-            'i18nStrings.valueText': 'Custom Value',
-            'i18nStrings.removeTokenButtonAriaLabel': `{token__operator, select, 
-              equals {Remove filter, {token__propertyKey} Custom equals {token__value}}
-              not_equals {Remove filter, {token__propertyKey} Custom does not equal {token__value}}
-              greater_than {Remove filter, {token__propertyKey} Custom greater than {token__value}}
-              greater_than_equal {Remove filter, {token__propertyKey} Custom greater than or equals {token__value}}
-              less_than {Remove filter, {token__propertyKey} Custom less than {token__value}}
-              less_than_equal {Remove filter, {token__propertyKey} Custom less than or equals {token__value}}
-              contains {Remove filter, {token__propertyKey} Custom contains {token__value}}
-              not_contains {Remove filter, {token__propertyKey} Custom does not contain {token__value}}
-              starts_with {Remove filter, {token__propertyKey} Custom starts with {token__value}}
-              not_starts_with {Remove filter, {token__propertyKey} Custom does not start with {token__value}}
-              other {}}`,
-          },
-        }}
-      >
-        <PropertyFilter
-          {...defaultProps}
-          tokenLimit={1}
-          query={{
-            operation: 'and',
-            tokens: [
-              { propertyKey: 'string', operator: '=', value: 'value1' },
-              { propertyKey: 'string', operator: '!=', value: 'value2' },
-              { propertyKey: 'string', operator: ':', value: 'value3' },
-              { propertyKey: 'string', operator: '!:', value: 'value4' },
-              { propertyKey: 'string', operator: '^', value: 'value5' },
-              { propertyKey: 'string', operator: '!^', value: 'value6' },
-              { propertyKey: 'range', operator: '>', value: '1' },
-              { propertyKey: 'range', operator: '<', value: '2' },
-              { propertyKey: 'range', operator: '>=', value: '3' },
-              { propertyKey: 'range', operator: '<=', value: '4' },
-            ],
-          }}
-          i18nStrings={{ filteringAriaLabel: 'your choice', filteringPlaceholder: 'Search' }}
-        />
-      </TestI18nProvider>
-    );
-    const wrapper = createWrapper(container).findPropertyFilter()!;
-    expect(wrapper.findRemoveAllButton()!.getElement()).toHaveTextContent('Custom Clear filters');
-    expect(wrapper.findTokenToggle()!.getElement()).toHaveTextContent('Custom Show more');
-    wrapper.findTokenToggle()!.click();
-    expect(wrapper.findTokenToggle()!.getElement()).toHaveTextContent('Custom Show fewer');
-
-    const getRemoveButton = (index: number) => wrapper.findTokens()[index].findRemoveButton().getElement();
-    expect(getRemoveButton(0)).toHaveAccessibleName('Remove filter, string Custom equals value1');
-    expect(getRemoveButton(1)).toHaveAccessibleName('Remove filter, string Custom does not equal value2');
-    expect(getRemoveButton(2)).toHaveAccessibleName('Remove filter, string Custom contains value3');
-    expect(getRemoveButton(3)).toHaveAccessibleName('Remove filter, string Custom does not contain value4');
-    expect(getRemoveButton(4)).toHaveAccessibleName('Remove filter, string Custom starts with value5');
-    expect(getRemoveButton(5)).toHaveAccessibleName('Remove filter, string Custom does not start with value6');
-    expect(getRemoveButton(6)).toHaveAccessibleName('Remove filter, range Custom greater than 1');
-    expect(getRemoveButton(7)).toHaveAccessibleName('Remove filter, range Custom less than 2');
-    expect(getRemoveButton(8)).toHaveAccessibleName('Remove filter, range Custom greater than or equals 3');
-    expect(getRemoveButton(9)).toHaveAccessibleName('Remove filter, range Custom less than or equals 4');
-
-    const tokenOperation = wrapper.findTokens()[1].findTokenOperation()!;
-    tokenOperation.openDropdown();
-    expect(tokenOperation.findDropdown()!.findOption(1)!.getElement()).toHaveTextContent('Custom and');
-    expect(tokenOperation.findDropdown()!.findOption(2)!.getElement()).toHaveTextContent('Custom or');
-
-    const [popoverContent, popover] = openTokenEditor(wrapper);
-    expect(popover.findHeader()!.getElement()).toHaveTextContent('Custom Edit filter');
-    expect(findPropertyField(popoverContent).findLabel()!.getElement()).toHaveTextContent('Custom Property');
-    expect(findOperatorField(popoverContent).findLabel()!.getElement()).toHaveTextContent('Custom Operator');
-    expect(findValueField(popoverContent).findLabel()!.getElement()).toHaveTextContent('Custom Value');
-    expect(findCancelButton(popoverContent).getElement()).toHaveTextContent('Custom Cancel');
-    expect(findSubmitButton(popoverContent).getElement()).toHaveTextContent('Custom Apply');
   });
 });
